Add the "force_destroy" argument to the S3 bucket configuration (disabled)
If the "force_destroy" argument is set to true, all objects (including
any locked objects) are deleted from the bucket when the bucket is
destroyed. Otherwise a "terraform destroy" command can fail (if the
bucket still contains objects).

As this configuration poses a risk of loosing data, the argument is
currently commented out.