Implement an HTTPS endpoint for SES notifications
Amazon SES notifies a SNS topic about bounce, complaint, and/or delivery
feedback. This change adds a SNS subscription to the SNS topic. The endpoint
is an HTTPS URI (default: https://example.com/ses-feedback-notifications).

The endpoint URI can be configured in the file variables.tf (root directory).

A random user name and a random password are generated and stored as a
secret in the Secrets Manager. These access credentials are sent as a basic
authentication in the request to the HTTPS endpoint.

Important note: The subscription must be confirmed. Amazon SNS will not send
requests to the endpoint until the subscription has been confirmed. See the
AWS documentation for further details.